crates/sel4-sync/abstractions: Introduce
