+++ Cargo.toml
@@ -136,6 +136,7 @@ members = [
     "crates/sel4-shared-ring-buffer/smoltcp",
     "crates/sel4-stack",
     "crates/sel4-sync",
+    "crates/sel4-sync/abstractions",
     "crates/sel4-synthetic-elf",
     "crates/sel4-test-harness",
     "crates/sel4/bitfield-ops",
