+++ crates/sel4-sync/abstractions/Cargo.toml
@@ -1,7 +1,7 @@
 #
 # Copyright 2023, Colias Group, LLC
 #
-# SPDX-License-Identifier: MIT
+# SPDX-License-Identifier: BSD-2-Clause
 #
 #
 # This file is generated from './Cargo.nix'. You can edit this file directly
@@ -10,14 +10,11 @@
 #
 
 [package]
-name = "sel4-sync"
+name = "sel4-sync-abstractions"
 version = "0.1.0"
 authors = ["Nick Spinale <nick.spinale@coliasgroup.com>"]
 edition = "2021"
-license = "MIT"
+license = "BSD-2-Clause"
 
 [dependencies]
 lock_api = "0.4.12"
-sel4 = { path = "../sel4" }
-sel4-immediate-sync-once-cell = { path = "../sel4-immediate-sync-once-cell" }
-sel4-sync-abstractions = { path = "abstractions" }
